+++ pocketutils/tools/numeric_tools.py
@@ -67,8 +67,8 @@ class NumericTools(BaseTools):
     @classmethod
     def slice_bounded(cls, arr: np.array, i: Optional[int], j: Optional[int]) -> np.array:
         """
-        Slices `arr[max(i,0), min(j, len(arr))`.
-        Converts `i` and `j` to int.
+        Slices ``arr[max(i,0), min(j, len(arr))``.
+        Converts ``i`` and ``j`` to int.
         """
         if i is None:
             i = 0