+++ tests/pocketutils/core/test_warning_utils.py
@@ -1,7 +1,5 @@
 import pytest
 
-from pocketutils.logging.warning_utils import *
-
 raises = pytest.raises