+++ tests/pocketutils/tools/test_filesys_tools.py
@@ -36,6 +36,20 @@ class TestFilesysTools:
         with pytest.raises(ParsingError):
             f(load("bad2.properties"))
 
+    def test_get_env_info(self):
+        data = FilesysTools.get_env_info(include_insecure=True)
+        assert len(data) > 20
+        assert "pid" in data
+        assert "disk_used" in data
+        assert "hostname" in data
+        assert "locale" in data
+
+    def test_list_imports(self):
+        data = FilesysTools.list_package_versions()
+        assert len(data) > 5
+        assert "orjson" in data
+        assert data["orjson"].startswith("3.")  # change when updated
+
 
 if __name__ == "__main__":
     pytest.main()